+++ common/src/main/java/lilypuree/metabolism/platform/services/MetabolismServerConfig.java
@@ -3,7 +3,10 @@ package lilypuree.metabolism.platform.services;
 public interface MetabolismServerConfig {
 
     boolean preciseFeedback();
+    
+    boolean disableHeat();
 
+    boolean convertResources();
 
 
     void reload();